foodlens-api/requirements.txt
fastapi
uvicorn
onnxruntime
pillow
numpy
opencv-python
python-multipart
pymongo
passlib[bcrypt]
python-jose